Add files via upload
The most significance of Machine Learning (ML) to use in Insurance Industry is to provide best 
possible predictions by building the model which processes the historical data of customer 
seamlessly to identify risk, claims and customer actions which further facilitates to make smart 
decisions and take appropriate actions. Insurance fraud affects enormously both financial 
aspects and everyday life. Frauds can reduce the trust in the company and degrades the growth 
of the company. This paper carries out a relative analysis of Insurance fraud detection using 
machine learning techniques to build machine learning models for prediction, which play a vital 
role in fraud detection, as it is implemented to extract and expose the hidden knowledge from
very large data.
The rest of this paper is organized as follows. Section 2 briefly discusses the methods to be 
carried out. Section 3 Result analysis. Section 4 is the Conclusion, Section 5 discusses future 
enhancements, Section 6 References used, and Section 7 Authors Profile.